// generated/src/app/api-models/lob-pre-exist-rule-dtl.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class LobPreExistRuleDtl {
  seqLbpecId: number;
  seqLbpecDtl: number;
  operator: string;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}